Use custom PeopleSync colors
The color palette is based on the Material design guidelines.

https://material.io/design/color/
https://material.io/design/color/#tools-for-picking-colors
https://www.materialpalette.com/

In contrast to DAVx5, we want a dark action bar, to get a white text on
dark background in the toolbar.

Change colorOnPrimary to white.
We want white on blue background, not black on blue.
This fixes the TabLayout in the About activity.