Trigger a sync request when the number of collection changes
The collection detection via DavService happens asynchronously in
AddressBooksSyncAdapterService. If a new collection is found during the
discovery process, the new address book would only be synced after yet
another sync schedule (the default is 4h).
To speed this up, trigger a manual sync request from DavService if the
number of collections changes.