Add license information for DAVx5 and related libraries
DAVx5 and its libraries are licensed under the GPL3.

We want to show in the About activity which components of DAVx5 we use
(including their version).
The AboutLibraries gradle plugin does not auto-detect the libraries we
pull in via git submodules, so we specify them manually via
about_libraries.xml.

We use the SPDX identifiers as the old name gpl_3_0 has issues.