Determine principal name on app upgrades
The username that is stored in the account settings is not necessarily
the principal name which we require to build the principal URL.
If old addressbooks exist, extract the principal name from the
addressbook URL.
We want to avoid hitting the network during the account migration.
So if we fail to determine the principal URL, simply delete the old
accounts. Leaving them around would break the app.