Improve background collection refreshes
Refactor DavService.refreshCollections() and make it public so it can be
used from AddressBooksSyncAdapterService.
We want to be able to trigger collection refreshes even if battery
optimisation has not been disabled.
A positive side-effect of this change is that
AddressBooksSyncAdapterService can now run the collection detection
synchronously, i.e. we only proceed with adding new address books or
deleting old ones after we have an up-to-date set of collections.
This means we no longer need to trigger a sync from within
refreshCollections().

Follow-up for commit fbe0b82701a5201b61fcbe07c410a721bc86d512.