RestrictionsProvider: map long values to int
Restriction entries don't have a Long type, only Integer [1].
Some settings values are of type long, such as "default_sync_interval".
Since we want to provide those via managed configurations, map long
values to int. This needs to be done carefully and only for values which
don't exceed int. In case of "default_sync_interval" it should be
unproblematic, as INT_MAX translations to an interval of 68 years and we
shouldn't exceed that.

[1] https://developer.android.com/reference/android/content/RestrictionEntry#TYPE_INTEGER