Fix About activity to include DAVx5 related data
We want to show in the About activity which components of DAVx5 we use
(including their version).
The AboutLibraries gradle plugin does not auto-detect the libraries we
pull in via git submodules, so we specify them manually via about.xml.

For dav4jvm, which is pulled in as an external dependency, we fix up the
meta data a little. See
https://forums.bitfire.at/topic/2324/could-not-get-the-name-for-com_gitlab_bitfireat__dav4jvm-using-com-gitlab-bitfireat-dav4jvm