PS-471 Do not expose explicit collection refreshes in the Account Activity
The PeopleSync app automatically synchronizes newly found address books,
and a sync also includes an implicit collection refresh.

This means, having a separate collection refresh action (which can
currently be triggered via the swipe down gesture, a separate floating
action button and a menu entry), is confusing.
If a new address book is found, it is shown as selected, but the actual
account is only created once a sync has been initiated.

So, do not expose the collection refresh action the Account Activity and
instead always use a full sync.

Only keep a single floating action button and the swipe gesture and drop
the menu entry to avoid a bit of duplication.