Remove CalDAV service detection in DavResourceFinder
This functionality was mainly left in to determine the email address for
the principal which then could be used as the account name. But since we
do not support CalDAV in the backend, this doesn't work anyway so better
remove it. This would need a different approach for a CardDAV only
backend.