Use managed configurations during account setup
Handle app restrictions directly in ManagedSettings

Using the SettingsProvider framework can lead to unexpected results,
especially when it concerns checking for the existence of a certain key.
Asking for the existence of a key can be misleading if the
DefaultsProvider answers that request.

So instead, setup ManageSettings to use app restrictions directly.
This means accessing (and caching) the app resrictions twice, in
ManagedSettings and RestrictionsProvider.